[misc] avoid useless truncation warning

+++ src/main/java/org/mariadb/jdbc/plugin/authentication/standard/ed25519/math/GroupElement.java
@@ -333,12 +333,12 @@ public class GroupElement implements Serializable {
     /* e[63] is between 0 and 7 */
     int carry = 0;
     for (i = 0; i < 63; i++) {
-      e[i] += carry;
+      e[i] += (byte) carry;
       carry = e[i] + 8;
       carry >>= 4;
-      e[i] -= carry << 4;
+      e[i] -= (byte)(carry << 4);
     }
-    e[63] += carry;
+    e[63] += (byte) carry;
     /* each e[i] is between -8 and 7 */
     return e;
   }
@@ -370,10 +370,10 @@ public class GroupElement implements Serializable {
           // Accumulate bits if possible
           if (r[i + b] != 0) {
             if (r[i] + (r[i + b] << b) <= 15) {
-              r[i] += r[i + b] << b;
+              r[i] += (byte) (r[i + b] << b);
               r[i + b] = 0;
             } else if (r[i] - (r[i + b] << b) >= -15) {
-              r[i] -= r[i + b] << b;
+              r[i] -= (byte) (r[i + b] << b);
               for (int k = i + b; k < 256; ++k) {
                 if (r[k] == 0) {
                   r[k] = 1;
